refactor: move mappers to util package for better organization
