+++ Backend/src/main/java/com/pg/mbti/util/mappers/UserMapper.java
@@ -1,4 +1,4 @@
-package com.pg.mbti.mappers;
+package com.pg.mbti.util.mappers;
 
 import com.pg.mbti.dto.UserProfileDto;
 import com.pg.mbti.entity.User;
